feat: 데이터 수집 및 정제 단계에서 팩토리 패턴 구상
