--- src/main/java/org/jun/saemangeum/process/application/dto/RefinedDataDTO.java
@@ -1,9 +1,0 @@
-package org.jun.saemangeum.process.application.dto;
-
-public record RefinedDataDTO(
-        String title,
-        String position,
-        String image,
-        String introduction
-) {
-}
