+++ src/main/java/org/jun/saemangeum/process/domain/dto/RefinedDataDTO.java
@@ -1,0 +1,12 @@
+package org.jun.saemangeum.process.domain.dto;
+
+import org.jun.saemangeum.global.persistence.domain.Category;
+
+public record RefinedDataDTO(
+        String title,
+        String position,
+        Category category,
+        String image,
+        String introduction
+) {
+}
